Abdullah branch

+++ lib/feature/home/createplayer/view/create_player_view.dart
@@ -20,18 +20,14 @@ class _CreatePlayerViewState extends CreatePlayerViewModel {
   }
 
   var textFieldController = TextEditingController();
-  final List<TextEditingController> _textEditingControllers = [
-    TextEditingController()
-  ];
+  final List<TextEditingController> _textEditingControllers = [TextEditingController()];
 
   Future<void> additemtoList(int index, int maxnumber) async {
     while (index < maxnumber) {
       _textEditingControllers.add(TextEditingController());
       if (Provider.of<Player>(context, listen: false).textValueisEmpty.length <
           Provider.of<GameSettingsModel>(context, listen: false).playerCount) {
-        Provider.of<Player>(context, listen: false)
-            .textValueisEmpty
-            .add(CreatePlayerViewModel.isEmpty);
+        Provider.of<Player>(context, listen: false).textValueisEmpty.add(CreatePlayerViewModel.isEmpty);
       }
 
       index++;
@@ -45,12 +41,11 @@ class _CreatePlayerViewState extends CreatePlayerViewModel {
     var screenHeight = screenInfo.size.height;
     var screenWidth = screenInfo.size.width;
     int? myId = Provider.of<Player>(context).id;
-    int? myScore = Provider.of<Player>(context).score;
+    double? myScore = Provider.of<Player>(context).score;
     int? myRank = Provider.of<Player>(context).rank;
     String? myImageString = Provider.of<Player>(context).image;
     // int userNumber = Provider.of<GameSettingsModel>(context).playerCount;
-    Map<dynamic, dynamic>? myPlayersMap =
-        Provider.of<Player>(context).playersMap;
+    Map<dynamic, dynamic>? myPlayersMap = Provider.of<Player>(context).playersMap;
     // Future<void> myEqualMap = Provider.of<Player>(context).equalMaps();
     // Future<String> writemaps(Map<dynamic, dynamic> writemap) async {
     //   return writemap.values.toString();
@@ -67,10 +62,7 @@ class _CreatePlayerViewState extends CreatePlayerViewModel {
         decoration: const BoxDecoration(
           gradient: RadialGradient(
             radius: 0.7,
-            colors: [
-              Color.fromRGBO(255, 149, 113, 1),
-              Color.fromRGBO(216, 91, 47, 1)
-            ],
+            colors: [Color.fromRGBO(255, 149, 113, 1), Color.fromRGBO(216, 91, 47, 1)],
           ),
         ),
         child: SizedBox(
@@ -110,6 +102,7 @@ class _CreatePlayerViewState extends CreatePlayerViewModel {
                           ? screenHeight / 1.6
                           : screenHeight / 1.4,
                   width: screenWidth / 1.15,
+
                   child: ListView.builder(
                     itemCount:
                         Provider.of<GameSettingsModel>(context).playerCount,
@@ -118,6 +111,7 @@ class _CreatePlayerViewState extends CreatePlayerViewModel {
                         index,
                         Provider.of<GameSettingsModel>(context).playerCount,
                       );
+
 
                       //burada player oluşturulabilir*-*-*-*-*-*-*-*-*
                       Provider.of<Player>(context).createPlayerfunc(
@@ -170,21 +164,18 @@ class _CreatePlayerViewState extends CreatePlayerViewModel {
       child: Row(
         children: [
           Padding(
-            padding:
-                EdgeInsets.only(left: MediaQuery.of(context).size.width / 20),
+            padding: EdgeInsets.only(left: MediaQuery.of(context).size.width / 20),
             child: SizedBox(
               width: MediaQuery.of(context).size.width / 5.2,
               height: MediaQuery.of(context).size.width / 5.2,
               child: Container(
                 decoration: BoxDecoration(
                     image: DecorationImage(
-                  image:
-                      AssetImage('assets/images/profiles/$numberforimages.png'),
+                  image: AssetImage('assets/images/profiles/$numberforimages.png'),
                   opacity: 1,
                   fit: BoxFit.fill,
                 )),
-                child: Provider.of<Player>(context, listen: false)
-                        .textValueisEmpty[numberforimages - 1]
+                child: Provider.of<Player>(context, listen: false).textValueisEmpty[numberforimages - 1]
                     ? const Center(
                         child: Text(
                           'HAZIR!',
@@ -207,8 +198,7 @@ class _CreatePlayerViewState extends CreatePlayerViewModel {
             ),
           ),
           Padding(
-            padding:
-                EdgeInsets.only(left: MediaQuery.of(context).size.width / 30),
+            padding: EdgeInsets.only(left: MediaQuery.of(context).size.width / 30),
             child: SizedBox(
                 height: 50,
                 width: 160,
@@ -219,8 +209,7 @@ class _CreatePlayerViewState extends CreatePlayerViewModel {
                   numberforimages,
                 )),
           ),
-          Provider.of<Player>(context, listen: false)
-                  .textValueisEmpty[numberforimages - 1]
+          Provider.of<Player>(context, listen: false).textValueisEmpty[numberforimages - 1]
               ? Padding(
                   padding: EdgeInsets.only(
                       left: MediaQuery.of(context).size.width / 40),
